refactor: moved UserRegisterUseCase to ifc
