+++ src/main/java/dev/chipichapa/memestore/usecase/ifc/UserRegisterUseCase.java
@@ -1,4 +1,4 @@
-package dev.chipichapa.memestore.usecase;
+package dev.chipichapa.memestore.usecase.ifc;
 
 import dev.chipichapa.memestore.dto.auth.RegisterRequest;
